Move main part of implementation into module
This allows making the whole module conditional on proc_macro during
no_std conversion, rather than cfg'ing every single line of code.
